% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{wearableAnomaly-package}
\alias{wearableAnomaly}
\alias{wearableAnomaly-package}
\title{wearableAnomaly: Tools for Wearable Device Anomaly Detection}
\description{
The package provides helpers to ingest, validate, resample, and analyse
wearable sensor signals ahead of anomaly detection workflows.
}
\author{
\strong{Maintainer}: Neo Kok \email{neokok@umich.edu}

}
